+++ src/interfaces/user.interface.ts
@@ -1,0 +1,7 @@
+export interface User {
+    id : string,
+    name? : string|null,
+    commands: number,
+    receivedWelcome: boolean,
+    admin: boolean,
+}
